% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resultUtils.R
\name{classifyResultFilesForMethod}
\alias{classifyResultFilesForMethod}
\title{method for classifying the files related to a certain method
according to their parameters}
\usage{
classifyResultFilesForMethod(method)
}
\arguments{
\item{method}{algorithm to analyze}
}
\value{
list with a number of entries equals to the number
of combinations of parameters for the arguments of the
method of interest
}
\description{
method for classifying the files related to a certain method
according to their parameters
}
