% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resultUtils.R
\name{composePatternName}
\alias{composePatternName}
\title{compose a pattern containing a method name and a combination
of arguments}
\usage{
composePatternName(method, combination)
}
\arguments{
\item{method}{method of interest}

\item{combination}{combination of parameters and values}
}
\value{
name of the file for the algorithm of interest and
the parameterization specified with the second argument
}
\description{
compose a pattern containing a method name and a combination
of arguments
}
