% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{computeoddsc}
\alias{computeoddsc}
\title{This function computes the odds P(nochange)/P(change) in two binomial
sequences}
\usage{
computeoddsc(n1, n2, x1, x2, s)
}
\arguments{
\item{n1}{sample size of first sequence}

\item{n2}{sample size of the second sequences}

\item{x1}{is the counting of 1s in first sequences}

\item{x2}{number of 1s in the second sequence}

\item{s}{the problem is solved with a Bayesian BDEu score with
equivalent sample size s}
}
\value{
odd (vector value with odds)
}
\description{
This function computes the odds P(nochange)/P(change) in two binomial
sequences
}
