% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{computepvalue}
\alias{computepvalue}
\title{compute p value for a test over four sequences (two serving as
references and two to check against)}
\usage{
computepvalue(x1, x2, x1n, x2n)
}
\arguments{
\item{x1}{number of 1s in reference serie}

\item{x2}{number of 0s in reference serie}

\item{x1n}{number of 1s in second serie}

\item{x2n}{number of 0s in second serie}
}
\value{
p value of statistic
}
\description{
compute p value for a test over four sequences (two serving as
references and two to check against)
}
