% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resultUtils.R
\name{computeStatistics}
\alias{computeStatistics}
\title{method for computing the global statistics about a concrete algorithm and
a certain combination of parameters}
\usage{
computeStatistics(method, arguments)
}
\arguments{
\item{arguments}{parameterizations to consider}

\item{algorithm}{of interest}
}
\value{
list with average and std. deviation for KL and time
}
\description{
method for computing the global statistics about a concrete algorithm and
a certain combination of parameters
}
