% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate1.R
\name{estimate1}
\alias{estimate1}
\title{Function that estimates probabilities from a string x
It contains a window of active values. It has a parameter n
for making the test: it compares the n last values with the n first
values of the windows (it must have a size greater or equal to 2*n)
if there are significant differences it removes the first n values
of the window and repeats the test.
alpha is the threshold for the odds}
\usage{
estimate1(x, n, alpha)
}
\arguments{
\item{x}{of data}

\item{n}{window size}

\item{alpha}{threshold for computing odds}
}
\value{
list with the estimations, list of the sample sizes,
and list of forgotten samples and values for ro
}
\description{
Function that estimates probabilities from a string x
It contains a window of active values. It has a parameter n
for making the test: it compares the n last values with the n first
values of the windows (it must have a size greater or equal to 2*n)
if there are significant differences it removes the first n values
of the window and repeats the test.
alpha is the threshold for the odds
}
