% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate10.R
\name{estimate10}
\alias{estimate10}
\title{Function that estimates probabilities from a string x
It returns a list with the estimations, the sample sizes, and the forgotten samples
forg is a vector of rhos. It considers all the rhos and selects
the rho with maximum likelihood in each case.}
\usage{
estimate10(x, forg, l)
}
\arguments{
\item{x}{stream to analyze}

\item{forg}{vector of values for rho parameter}

\item{l}{length to consider}
}
\value{
list with data, estimations and values of s and ro
}
\description{
Function that estimates probabilities from a string x
It returns a list with the estimations, the sample sizes, and the forgotten samples
forg is a vector of rhos. It considers all the rhos and selects
the rho with maximum likelihood in each case.
}
