% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate11-ADWIN.R
\name{estimate11}
\alias{estimate11}
\title{Base estimation method termed ADWIN}
\usage{
estimate11(x, delta)
}
\arguments{
\item{x}{stream to analyze}

\item{delta}{parameter}
}
\value{
list with data, estimations and values of s and ro
}
\description{
Base estimation method termed ADWIN
}
