% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate12.R
\name{estimate12}
\alias{estimate12}
\title{Function that estimates probabilities from a string x
It returns a list with the estimations, the sample sizes, and the
forgotten samples
It contains a window of active values.
It is similar to estimate7, but now it carries out a chisquared test}
\usage{
estimate12(x, n, alpha1, alpha2)
}
\arguments{
\item{x}{stream to analyze}

\item{n}{size of active window}

\item{alpha1}{parameter for test}

\item{alpha2}{parameter for test}
}
\value{
list with data, estimations and values of s and ro
}
\description{
Function that estimates probabilities from a string x
It returns a list with the estimations, the sample sizes, and the
forgotten samples
It contains a window of active values.
It is similar to estimate7, but now it carries out a chisquared test
}
