% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate13.R
\name{estimate13}
\alias{estimate13}
\title{Function that estimates probabilities from a string x
It returns a list with the estimations, the sample sizes, and the
forgotten samples
It contains a window of active values.
It is similar to estimate1 and estimate5, but now the window of active
values is divided in two
parts if its length is greater than n: one with the last n values and other
with the rest of values}
\usage{
estimate13(x, n, alpha1, alpha2)
}
\arguments{
\item{x}{stream to analyze}

\item{n}{size of active window}

\item{alpha1}{parameter for test}

\item{alpha2}{parameter for test}
}
\value{
list with data, estimations and values of s and ro
}
\description{
Function that estimates probabilities from a string x
It returns a list with the estimations, the sample sizes, and the
forgotten samples
It contains a window of active values.
It is similar to estimate1 and estimate5, but now the window of active
values is divided in two
parts if its length is greater than n: one with the last n values and other
with the rest of values
}
