% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate14.R
\name{estimate14}
\alias{estimate14}
\title{Function that estimates probabilities from a string x
It returns a list with the estimations, the sample sizes, and the
forgotten samples
It contains a window of active values.
It is similar to estimate1 and estimate5, but now the window of active
values is divided in two
parts if its length is greater than n: one with the last n values and
other with the rest of values}
\usage{
estimate14(x, n, delta)
}
\arguments{
\item{x}{stream to analyze}

\item{n}{size of active window}

\item{delta}{parameter for test}
}
\value{
list with data, estimations and values of s and ro
}
\description{
Function that estimates probabilities from a string x
It returns a list with the estimations, the sample sizes, and the
forgotten samples
It contains a window of active values.
It is similar to estimate1 and estimate5, but now the window of active
values is divided in two
parts if its length is greater than n: one with the last n values and
other with the rest of values
}
