% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate15-SWMTF.R
\name{estimate15}
\alias{estimate15}
\title{Estimation method termed SWMTF}
\usage{
estimate15(x, alpha)
}
\arguments{
\item{x}{stream to analyze}

\item{alpha}{parameter for test}
}
\value{
list with data, estimations, values of s and ro
}
\description{
Estimation method termed SWMTF
}
