% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate16.R
\name{estimate16}
\alias{estimate16}
\title{Estimation function}
\usage{
estimate16(x, alpha1, alpha2)
}
\arguments{
\item{x}{stream to analyze}

\item{alpha1}{value for test}

\item{alpha2}{value for test}
}
\value{
list with data, estimations, values for s, fg and ro
}
\description{
Estimation function
}
