% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate18.R
\name{estimate18}
\alias{estimate18}
\title{Estimation method}
\usage{
estimate18(x, alpha1, alpha2)
}
\arguments{
\item{x}{stream to analyze}

\item{alpha1}{value for test}

\item{alpha2}{value for test}
}
\value{
list with data, estimations, values for s, fg and ro
}
\description{
Estimation method
}
