% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate2.R
\name{estimate2}
\alias{estimate2}
\title{Function that estimates probabilities from a string x
It returns a list with the estimations, the sample sizes, the forgotten
samples, and the forgetting coefficients
It makes statistical tests (goodness of fit) of the n1 last samples with
respect to the estimations done on step i-n1. If test is meaningfull, then
it multiplies the past samples by a coefficient rho
m is a value such that the p-value is squared to m}
\usage{
estimate2(x, n1, m)
}
\arguments{
\item{x}{stream to analyze}

\item{n1}{number of recent samples to consider}

\item{m}{p-value will be squared to m}
}
\description{
Function that estimates probabilities from a string x
It returns a list with the estimations, the sample sizes, the forgotten
samples, and the forgetting coefficients
It makes statistical tests (goodness of fit) of the n1 last samples with
respect to the estimations done on step i-n1. If test is meaningfull, then
it multiplies the past samples by a coefficient rho
m is a value such that the p-value is squared to m
}
