% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate20.R
\name{estimate20}
\alias{estimate20}
\title{Estimation method}
\usage{
estimate20(x, n, alpha)
}
\arguments{
\item{x}{stream to analyze}

\item{n}{size of active window}

\item{alpha}{value for test}
}
\value{
list with data, estimations, values for s, fg and ro
}
\description{
Estimation method
}
