% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate21-SWMTFIn.R
\name{estimate21}
\alias{estimate21}
\title{Estimation method termed SWMTFIn}
\usage{
estimate21(x, n, alpha)
}
\arguments{
\item{x}{stream to analyze}

\item{n}{size of active window}

\item{alpha}{value for test}
}
\value{
list with data, estimations, values for s, fg and ro
}
\description{
Estimation method termed SWMTFIn
}
