% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate23-SWMTB.R
\name{estimate23}
\alias{estimate23}
\title{Estimation method named SWMTB}
\usage{
estimate23(x, alpha)
}
\arguments{
\item{x}{stream to analyze}

\item{alpha}{value for test}
}
\value{
list with data, estimations, values for s, fg and ro
}
\description{
Estimation method named SWMTB
}
