% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate24-SWMTBIn.R
\name{estimate24}
\alias{estimate24}
\title{Estimation method named SWMTBIn}
\usage{
estimate24(x, n, alpha)
}
\arguments{
\item{x}{stream to analyze}

\item{n}{size of active window}

\item{alpha}{value for test}
}
\value{
list with data, estimations, values for s, fg and ro
}
\description{
Estimation method named SWMTBIn
}
