% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate26.R
\name{estimate26}
\alias{estimate26}
\title{Estimation method}
\usage{
estimate26(x, n, alpha)
}
\arguments{
\item{x}{stream to analyze}

\item{n}{size of active window}

\item{alpha}{value for test}
}
\value{
list with data, estimations, values for s, fg and ro
}
\description{
Estimation method
}
