% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate29.R
\name{estimate29}
\alias{estimate29}
\title{Function that estimates the probabilities with a  forgetting factor
which is computed by a Bayesian BDEu score of the probability of change
in the last n cases (dividing these cases in two equal parts)}
\usage{
estimate29(x, n, alpha)
}
\arguments{
\item{x}{stream to analyze}

\item{n}{size of active window}

\item{alpha}{value for test}
}
\value{
list with data, estimations, values for s, fg and ro
}
\description{
Function that estimates the probabilities with a  forgetting factor
which is computed by a Bayesian BDEu score of the probability of change
in the last n cases (dividing these cases in two equal parts)
}
