% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate3-FF.R
\name{estimate3}
\alias{estimate3}
\title{Function that estimates the probabilities with a constant forgeting
factor ro}
\usage{
estimate3(x, ro)
}
\arguments{
\item{x}{stream to analyze}

\item{ro}{forgetting factor}
}
\value{
list with data, estimations, values for s, fg and ro
}
\description{
Function that estimates the probabilities with a constant forgeting
factor ro
}
