% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate34-BAF.R
\name{estimate34}
\alias{estimate34}
\title{Estimation method named BAF}
\usage{
estimate34(x, n, alpha)
}
\arguments{
\item{x}{stream to analyze}

\item{n}{size of active window}

\item{alpha}{value for test}
}
\value{
list with data, estimations, values for s, fg and ro
}
\description{
Estimation method named BAF
}
