% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate35-BFV1.R
\name{estimate35}
\alias{estimate35}
\title{Estimation method named BFV1}
\usage{
estimate35(x, n, alpha1, alpha2, k)
}
\arguments{
\item{x}{stream to analyze}

\item{n}{size of active window}

\item{alpha1}{value for test}

\item{alpha2}{value for test}

\item{k}{parameter value}
}
\value{
list with data, estimations, values for s, fg and ro
}
\description{
Estimation method named BFV1
}
