% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate36-BFV2.R
\name{estimate36}
\alias{estimate36}
\title{Estimation method named BFV2}
\usage{
estimate36(x, n, alpha1, alpha2, k)
}
\arguments{
\item{x}{stream to analyze}

\item{n}{size of active window}

\item{alpha1}{value for test}

\item{alpha2}{value for test}

\item{k}{parameter value}
}
\value{
list with data, estimations, values for s, fg and ro
}
\description{
Estimation method named BFV2
}
