% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate4-FW.R
\name{estimate4}
\alias{estimate4}
\title{Function that estimates probabilities from a string x
It returns a list with the estimations, the sample sizes, and the forgotten
samples
It contains a window of active values with fixex size n}
\usage{
estimate4(x, n)
}
\arguments{
\item{x}{stream to analyze}

\item{n}{size of active window}
}
\value{
list with data, estimations, values of s and ro
}
\description{
Function that estimates probabilities from a string x
It returns a list with the estimations, the sample sizes, and the forgotten
samples
It contains a window of active values with fixex size n
}
