% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate6.R
\name{estimate6}
\alias{estimate6}
\title{Function that estimates probabilities from a string x
It returns a list with the estimations, the sample sizes, and the
forgotten samples
It contains a window of active values.
It is similar to estimate1, but now the window of active
values is divided by two, if the total size is greater
than n}
\usage{
estimate6(x, n, alpha)
}
\arguments{
\item{x}{stream of data to analyze}

\item{n}{size of active window}

\item{alpha}{for tests}
}
\value{
list of data, estimations, values of s and ro
}
\description{
Function that estimates probabilities from a string x
It returns a list with the estimations, the sample sizes, and the
forgotten samples
It contains a window of active values.
It is similar to estimate1, but now the window of active
values is divided by two, if the total size is greater
than n
}
