% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate7-SWB.R
\name{estimate7}
\alias{estimate7}
\title{Function that estimates probabilities from a string x
It returns a list with the estimations, the sample sizes, and the
forgotten samples
It contains a window of active values.
It is similar to estimate1 and estimate5, but now the window of active
values is divided in two
parts if its length is greater than n: one with the last n values and other
with the rest of values}
\usage{
estimate7(x, n, alpha)
}
\arguments{
\item{x}{stream to analyze}

\item{n}{size of active window}

\item{alpha}{value for tests}
}
\value{
list with data, estimations and values of s and ro
}
\description{
Function that estimates probabilities from a string x
It returns a list with the estimations, the sample sizes, and the
forgotten samples
It contains a window of active values.
It is similar to estimate1 and estimate5, but now the window of active
values is divided in two
parts if its length is greater than n: one with the last n values and other
with the rest of values
}
