% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate8-SWF.R
\name{estimate8}
\alias{estimate8}
\title{Function that estimates probabilities from a string x
It returns a list with the estimations, the sample sizes, and the
forgotten samples
It contains a window of active values.
It is similar to estimate7, but now it carries out a chisquared test}
\usage{
estimate8(x, n, alpha)
}
\arguments{
\item{x}{stream to analyze}

\item{n}{size of active window}

\item{alpha}{value for tests}
}
\value{
list with data, estimations and values of s and ro
}
\description{
Function that estimates probabilities from a string x
It returns a list with the estimations, the sample sizes, and the
forgotten samples
It contains a window of active values.
It is similar to estimate7, but now it carries out a chisquared test
}
