% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate9.R
\name{estimate9}
\alias{estimate9}
\title{Function that estimates probabilities from a string x
It returns a list with the estimations, the sample sizes, and the
forgotten samples
forg is a vector of rhos. It considers all the rhos and makes an average
of the estimations with the different rhos weighted by the likelihood of
the data given each rho}
\usage{
estimate9(x, forg, l)
}
\arguments{
\item{x}{stream to analyze}

\item{forg}{vector of rho values}

\item{l}{length to consider}
}
\value{
list with data, estimations and values of s and ro
}
\description{
Function that estimates probabilities from a string x
It returns a list with the estimations, the sample sizes, and the
forgotten samples
forg is a vector of rhos. It considers all the rhos and makes an average
of the estimations with the different rhos weighted by the likelihood of
the data given each rho
}
