% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execution.R
\name{experiment}
\alias{experiment}
\title{read a file containing the description of an experiment and
executes it}
\usage{
experiment(name)
}
\arguments{
\item{name}{of the file describing the test to perform}
}
\description{
read a file containing the description of an experiment and
executes it
}
