% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fiadb.r
\name{fia_connect}
\alias{fia_connect}
\title{Establish connection to FIADB database}
\usage{
fia_connect(fiadb)
}
\arguments{
\item{fiadb}{-- fully qualified file name of an FIADB SQLite database}
}
\value{
Returns an connection to the SQLite DB.
}
\description{
Establish connection to FIADB database
}
\examples{
# connect to Maine FIADB database
# fia.connect( "c:/temp/ME.db" )
}
