% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resultUtils.R
\name{generateGlobalMatrixForTest}
\alias{generateGlobalMatrixForTest}
\title{generates a global matrix of results for a set of methods and
parameterizations}
\usage{
generateGlobalMatrixForTest(methods, argumentsForMethod)
}
\arguments{
\item{methods}{list of algorithms of interest}

\item{argumentsForMethods}{list of arguments for selected methods}
}
\value{
list with kl and time values
}
\description{
generates a global matrix of results for a set of methods and
parameterizations
}
