% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resultUtils.R
\name{generateLatexReport}
\alias{generateLatexReport}
\title{method for generating a latex report with the results for a certaint
method of interest}
\usage{
generateLatexReport(method)
}
\arguments{
\item{method}{algorithm of interest}
}
\description{
method for generating a latex report with the results for a certaint
method of interest
}
