% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resultUtils.R
\name{generateMethodMatrixForTest}
\alias{generateMethodMatrixForTest}
\title{generate a matrix with the results of the algorithm
given by method and the arguments indicated by the
second argument. It will produce a matrix with a colum
(with the name of the algorithm) and the rows will
be the results of the runs executed for it}
\usage{
generateMethodMatrixForTest(method, arguments)
}
\arguments{
\item{method}{algorithm of interest}

\item{arguments}{parameterization of interest}
}
\value{
list with KL and time values
}
\description{
generate a matrix with the results of the algorithm
given by method and the arguments indicated by the
second argument. It will produce a matrix with a colum
(with the name of the algorithm) and the rows will
be the results of the runs executed for it
}
