% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resultUtils.R
\name{generatePlot}
\alias{generatePlot}
\title{generate a plot for a method with the results stored in a certain
file}
\usage{
generatePlot(method, filename)
}
\arguments{
\item{method}{method of interest}

\item{filename}{name of the file to analyze}
}
\description{
generate a plot for a method with the results stored in a certain
file
}
