% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resultUtils.R
\name{generateTable}
\alias{generateTable}
\title{auxiliar method for generating a table with the results contained in the
matrix for a certain set of cells}
\usage{
generateTable(method, matrix, argsColumns, resultColumns, firstIndex,
  lastIndex)
}
\arguments{
\item{method}{algorithm of interest}

\item{matrix}{matrix with results}

\item{argsColumns}{columns of interest in matrix}

\item{resultColumns}{columns of results to consider}
}
\description{
auxiliar method for generating a table with the results contained in the
matrix for a certain set of cells
}
