% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fiadb.r
\name{get_fia}
\alias{get_fia}
\title{Get a STAND_CN's data from an FIADB database for growing with ACD.}
\usage{
get_fia(fiadb.con, stand.condition.code)
}
\arguments{
\item{stand.condition.code}{-- FIA stand condition code (STAND_CN) to retrieve}

\item{fiadbcon}{-- Connection to an FIADB database (see `fia.connect()`)}
}
\value{
Returns a list with two members if successful:
\itemize{
    \item stand : a data.frame with stand/plot level data from the STAND_CN
    \item tree : a data.frame with tree records from the STAND_CN
}

If an error occurs, the function returns `NULL`.
}
\description{
Get a STAND_CN's data from an FIADB database for growing with ACD.
}
\examples{
# connect to Oregon FIADB database
# or.db <- fia.connect( "c:/temp/OR.db" )
# fiadata <- get.fia( or.db, 1234567890 )

}
