% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resultUtils.R
\name{getArgumentNames}
\alias{getArgumentNames}
\title{get the names of the arguments composing the name of a file with
results}
\usage{
getArgumentNames(fileName)
}
\arguments{
\item{fileName}{name of the file of interest}
}
\value{
list of argument names
}
\description{
get the names of the arguments composing the name of a file with
results
}
