% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resultUtils.R
\name{getArgumentValues}
\alias{getArgumentValues}
\title{get the values of the arguments composing the name of a file with
results}
\usage{
getArgumentValues(fileName)
}
\arguments{
\item{fileName}{name of the file to consider}
}
\value{
list of values for the arguments
}
\description{
get the values of the arguments composing the name of a file with
results
}
