% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resultUtils.R
\name{getFilesForMethod}
\alias{getFilesForMethod}
\title{get the complete list of values for a certain method}
\usage{
getFilesForMethod(method)
}
\arguments{
\item{method}{method of interest}
}
\value{
list of files matching the arguments
}
\description{
get the complete list of values for a certain method
}
