% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resultUtils.R
\name{getResultFiles}
\alias{getResultFiles}
\title{gets the result files for a certain method and a named vector of
arguments.}
\usage{
getResultFiles(method, arguments)
}
\arguments{
\item{method}{method of interest}

\item{arguments}{combinations of arguments of interest}
}
\value{
list of files matching the values of the arguments
}
\description{
gets the result files for a certain method and a named vector of
arguments.
}
