% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resultUtils.R
\name{getResultsDataFrame}
\alias{getResultsDataFrame}
\title{gets the results as a data frame}
\usage{
getResultsDataFrame(method, referenceCriteria, toSelect)

getResultsDataFrame(method, referenceCriteria, toSelect)
}
\arguments{
\item{method}{algorithm of interest}

\item{referenceCriteria}{criteria used for selection}

\item{toSelect}{number of rows to select}

\item{method}{algorithm of interest}
}
\value{
data frame with results

data frame with results
}
\description{
gets the results as a data frame

gets the results as a data frame using a certain criteria
for selecting results
}
