% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resultUtils.R
\name{getResultsMatrix}
\alias{getResultsMatrix}
\title{gets the matrix of results for a given algorithm}
\usage{
getResultsMatrix(method)
}
\arguments{
\item{method}{algorithm of interest}
}
\value{
matrix with result for the files of results related to a
certain method. The columns of the matrix will contain information
about klMean, klStdev, timeMean and timeStdev
}
\description{
gets the matrix of results for a given algorithm
}
