% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resultUtils.R
\name{getRowMaxKL}
\alias{getRowMaxKL}
\title{gets the row with the max value for KL}
\usage{
getRowMaxKL(matrix)
}
\arguments{
\item{matrix}{matrix with results}
}
\description{
gets the row with the max value for KL
}
