% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resultUtils.R
\name{getRowMaxTime}
\alias{getRowMaxTime}
\title{gets the row with the max value for computation time}
\usage{
getRowMaxTime(matrix)
}
\arguments{
\item{matrix}{matrix with results}
}
\description{
gets the row with the max value for computation time
}
