% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resultUtils.R
\name{getRowMinKL}
\alias{getRowMinKL}
\title{gets the row with the min value for KL}
\usage{
getRowMinKL(matrix)
}
\arguments{
\item{matrix}{matrix with results}
}
\description{
gets the row with the min value for KL
}
