% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resultUtils.R
\name{getRowMinTime}
\alias{getRowMinTime}
\title{gets the row with the min value for computation time}
\usage{
getRowMinTime(matrix)
}
\arguments{
\item{matrix}{matrix with results}
}
\description{
gets the row with the min value for computation time
}
