% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resultUtils.R
\name{getValuesForArgument}
\alias{getValuesForArgument}
\title{get a complete vector of values for a certain argument in a
collection of files containing results}
\usage{
getValuesForArgument(fileNames, argumentName)
}
\arguments{
\item{fileNames}{list of files to analyze}

\item{argumentName}{name of the argument to look for}
}
\value{
vector of values for the argument considered in
the list of file names
}
\description{
get a complete vector of values for a certain argument in a
collection of files containing results
}
