% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resultUtils.R
\name{getValuesForLastArgument}
\alias{getValuesForLastArgument}
\title{gets the values for the last argument from matrix}
\usage{
getValuesForLastArgument(matrix)
}
\arguments{
\item{matrix}{matrix with results}
}
\description{
gets the values for the last argument from matrix
}
