% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{grow_acd}
\alias{grow_acd}
\title{grow_acd() : grow tree list with Acadian Variant of FVS optionally using metric or imperial units.}
\usage{
grow_acd(
  periods,
  Region,
  year,
  units,
  csi,
  elev,
  cdef,
  use_sbw,
  use_hw,
  use_thin,
  use_ingrowth,
  cut_point,
  MinDBH,
  plot_id,
  tree_id,
  spp,
  dbh,
  ht,
  expf,
  cr,
  form,
  risk
)
}
\arguments{
\item{periods}{: Integer | Number of periods to project the tree list}

\item{Region}{: String  | Region code: 'ME' (Maine), 'NB' (New Brunswick)}

\item{year}{: Integer | Current year or age of stand}

\item{units}{: Integer | 0 = metric, 1 = imperial}

\item{csi}{: Numeric | Site Index (meters or feet)}

\item{elev}{: Numeric | Elevation (meters or feet)}

\item{cdef}{: Numeric | Cumulative Defoliation Percentage (between 0 and 1)}

\item{use_sbw}{: Integer | Use Spruce Budworm Modifiers (0 = No, 1 = Yes)}

\item{use_hw}{: Integer | Use Hardwood Risk and Form Modifiers (0 = No, 1 = Yes)}

\item{use_thin}{: Integer | Use Thinning Modifiers (0 = No, 1 = Yes)}

\item{use_ingrowth}{: Integer | Use ingrowth functionality}

\item{cut_point}{: Numeric | ingrowth probability threshold}

\item{MinDBH}{: Numeric | Minimum DBH for ingrowth trees (cm or inches)}

\item{plot_id}{: vector  | Plot ID}

\item{tree_id}{: vector  | Tree ID}

\item{spp}{: vector  | FIA Numeric Species Code}

\item{dbh}{: vector  | diameter at breast height (cm or inches)}

\item{ht}{: vector  | total height (meters or feet)}

\item{expf}{: vector  | expansion factor, trees per hectare or per acre}

\item{cr}{: vector  | crown ratio (between 0 and 1)}

\item{form}{: vector  | Form Code (Castle et al. 2017) 1 through 8}

\item{risk}{: vector  | Risk Code (Castle et al. 2017) 1 through 4}
}
\value{
Returns a data.frame with the grown tree list in following variables:
\itemize{
\item plot.id
\item tree.id
\item species
\item dbh
\item ht
\item expf
\item cr
}
}
\description{
Projects the tree list using the Acadian Variant of FVS with Metric Units.

FIA species codes are defined here: https://www.fia.fs.usda.gov/program-features/urban/docs/Appendix\%203a\%20CORE\%20Species\%20List.pdf.
}
