% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fiadb.r
\name{grow_acd_fia}
\alias{grow_acd_fia}
\title{grow_acd_fia() : grow an FIA tree list with Acadian Variant of FVS optionally using metric or imperial units.}
\usage{
grow_acd_fia(
  db,
  stand_cn,
  periods,
  cdef = 0,
  use_sbw = F,
  use_hw = F,
  use_thin = F,
  use_ingrowth = F,
  cut_point = 0.5,
  MinDBH = 1.18,
  form = NA,
  risk = NA,
  csi_default = 52.5
)
}
\arguments{
\item{db}{: SQLite DB connection}

\item{stand_cn}{: FIA STAND_CN identifier}

\item{periods}{: Number of periods to project the tree list}

\item{cdef}{: Cumulative Defoliation Percentage (between 0 and 1) (default = 0)}

\item{use_sbw}{: Use Spruce Budworm Modifiers (F = No (default)), T = Yes)}

\item{use_hw}{: Use Hardwood Risk and Form Modifiers (F = No (default)), T = Yes)}

\item{use_thin}{: Use Thinning Modifiers (F = No (default)), T = Yes)}

\item{use_ingrowth}{:  Use ingrowth functionality (F = No (default)), T = Yes)}

\item{cut_point}{: Ingrowth probability threshold (default = 0.5)}

\item{MinDBH}{: Minimum DBH for ingrowth trees (cm or inches) (default = 1.18in)}

\item{form}{: Optional Form Code (Castle et al. 2017) 1 through 8 (default = NA)}

\item{risk}{: Optional Risk Code (Castle et al. 2017) 1 through 4 (default = NA)}

\item{csi_default}{: Default climate site index used if site is missing (feet) (defaults to 52.5ft)}
}
\value{
Returns a data.frame with the grown tree list in following variables (all values in imperial units):
\itemize{
    \item STAND_CN
    \item PLOT_ID
    \item TREE_ID
    \item SPECIES
    \item DBH
    \item HT
    \item TPA
    \item CR
}
}
\description{
Extracts an FIA condition from an open FIA DB data base connection (see \link{fia_connect}). Projects the extracted
tree list using the Acadian Variant of FVS.
}
\examples{
# open FIA DB connection
db <- fia_connect( "C:/Models/FIA/WV/SQLite_FIADB_WV.db" )

# project the FIA condition 7340111010661 10 years
grown.plot <- grow_acd_fia( db, 7340111010661, 10 )

}
