% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{kl}
\alias{kl}
\title{computes Kullbac-Leibler distance}
\usage{
kl(y, z)
}
\arguments{
\item{y}{the vector of estimated probabilities}

\item{z}{the vector of true probabilities}
}
\value{
computed value
}
\description{
computes Kullbac-Leibler distance
}
