% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{kld}
\alias{kld}
\title{Computes Kullback-Leibler distance}
\usage{
kld(y1, y2, z)
}
\arguments{
\item{y1}{a vector with the estimated probabilities}

\item{y2}{a vector with the true probabilities}

\item{z}{a vector of observations}
}
\value{
KL computation
}
\description{
Computes Kullback-Leibler distance
}
