% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{kls}
\alias{kls}
\title{Computes KLs distance}
\usage{
kls(z1, z2)
}
\arguments{
\item{z1}{the true probability of 1}

\item{z2}{the estimated probability of 1}
}
\value{
res value
}
\description{
Computes KLs distance
}
