% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{logl}
\alias{logl}
\title{Computes log-likelihood
TODO: is the same code as kls}
\usage{
logl(z1, z2)
}
\arguments{
\item{z1}{the true probability of 1}

\item{z2}{the estimated probability of 1}
}
\value{
res value
}
\description{
Computes log-likelihood
TODO: is the same code as kls
}
