% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resultUtils.R
\name{makeDocumentPreamble}
\alias{makeDocumentPreamble}
\title{compuses the preamble of the latex document with the
results}
\usage{
makeDocumentPreamble(method)
}
\arguments{
\item{method}{algorithm of interest}
}
\description{
compuses the preamble of the latex document with the
results
}
