% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resultUtils.R
\name{makeLabelWithParameters}
\alias{makeLabelWithParameters}
\title{auxiliar method forpreparing the label to include in a graphic from the
name of a file}
\usage{
makeLabelWithParameters(filename)
}
\arguments{
\item{filename}{name of the file of interest}
}
\description{
auxiliar method forpreparing the label to include in a graphic from the
name of a file
}
