% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resultUtils.R
\name{makeString}
\alias{makeString}
\title{compounds a string with the method and the concatenation of method,
argument names and corresponding values. Tha args are specified throw
selecting from the matrix of results the columns of a row containing
the values}
\usage{
makeString(method, args)
}
\arguments{
\item{method}{algorithm of interest}

\item{args}{argument names and values}
}
\description{
compounds a string with the method and the concatenation of method,
argument names and corresponding values. Tha args are specified throw
selecting from the matrix of results the columns of a row containing
the values
}
