% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resultUtils.R
\name{plotResultForCombination}
\alias{plotResultForCombination}
\title{plots the results considering a list of result files and taking
into account a certain parameter}
\usage{
plotResultForCombination(listOfFiles, valueName)
}
\description{
plots the results considering a list of result files and taking
into account a certain parameter
}
