% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resultUtils.R
\name{plotSeparatedSerieResults}
\alias{plotSeparatedSerieResults}
\title{generate plots for the results stored in the matrix and grouping
them respect to a given parameter}
\usage{
plotSeparatedSerieResults(matrix, groupingByAttrib, xAttrib, yAttrib)
}
\arguments{
\item{matrix}{matrixWithResults}

\item{groupingByAttrib}{attribute used for grouping results}

\item{xAttrib}{value for X axis}

\item{yAttrib}{value for Y axis}
}
\description{
generate plots for the results stored in the matrix and grouping
them respect to a given parameter
}
