% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resultUtils.R
\name{plotSerieResults}
\alias{plotSerieResults}
\title{plots the results for a certain method accordinf to a given
parameter}
\usage{
plotSerieResults(method, valueName)
}
\arguments{
\item{method}{algorithm of interest}

\item{valueName}{name of the parameter of interest}
}
\description{
plots the results for a certain method accordinf to a given
parameter
}
