% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resultUtils.R
\name{plotSerieResultsGgplot}
\alias{plotSerieResultsGgplot}
\title{genrates ggplot graphics for a matrix of results}
\usage{
plotSerieResultsGgplot(matrix, xAttrib, yAttrib)
}
\arguments{
\item{matrix}{matric with results}

\item{xAttrib}{value for X axis}

\item{yAttrib}{value for Y axis}
}
\description{
genrates ggplot graphics for a matrix of results
}
