% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resultUtils.R
\name{processResultFiles}
\alias{processResultFiles}
\title{process a list of files reading their content and computing KL and time
statistisc. The result is a list with entries for: arguments, klMean,
klStedev, timeMean and timeStdev}
\usage{
processResultFiles(listOfFiles)
}
\arguments{
\item{list}{of files}
}
\value{
list with entries for arguments, klMean, klStdev, timeMean and
timeStdev
}
\description{
process a list of files reading their content and computing KL and time
statistisc. The result is a list with entries for: arguments, klMean,
klStedev, timeMean and timeStdev
}
