% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resultUtils.R
\name{processResultFilesMethod}
\alias{processResultFilesMethod}
\title{method for processing the files for a given method}
\usage{
processResultFilesMethod(method)
}
\arguments{
\item{method}{algorithm of interest}
}
\description{
method for processing the files for a given method
}
