% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resultUtils.R
\name{removeExtensionOfFileNamesList}
\alias{removeExtensionOfFileNamesList}
\title{remove the extension of a list of files with results}
\usage{
removeExtensionOfFileNamesList(fileNamesList)
}
\arguments{
\item{fileNamesList}{list of file names}
}
\value{
list of files after removing extensions
}
\description{
remove the extension of a list of files with results
}
