% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execution.R
\name{saveResult}
\alias{saveResult}
\title{Saves the result of a experiment into a file in order to obtain
posterior data about computation}
\usage{
saveResult(method, iteration, args, results, time, realp, kl)
}
\arguments{
\item{method}{id of method employed}

\item{iteration}{related to the results}

\item{args}{arguments used}

\item{results}{produces by estimation method}

\item{time}{execution time}

\item{realp}{real value of p parameter}

\item{kl}{value of kullback-leibler distance between estimation and
real value of the parameter}
}
\description{
Saves the result of a experiment into a file in order to obtain
posterior data about computation
}
