% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execution.R
\name{sexp}
\alias{sexp}
\title{Function that calls to the appropriate estimation procedure for a set
of parameters}
\usage{
sexp(x, param, rp, iteration)
}
\arguments{
\item{x}{stream of data}

\item{param}{parameters for the experiment}

\item{rp}{value of p parameter for each sample}

\item{iteration}{number of iteration}
}
\description{
Function that calls to the appropriate estimation procedure for a set
of parameters
}
