% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate11-ADWIN.R
\name{sexp11}
\alias{sexp11}
\title{Function that calls to estimate11 for a set of parameters}
\usage{
sexp11(x, param, rp, iteration)
}
\arguments{
\item{x}{stream of data}

\item{param}{parameters for the experiment}

\item{rp}{real value of p parameter for each sample}

\item{iteration}{number of iteration}
}
\value{
list with results of execution
}
\description{
Function that calls to estimate11 for a set of parameters
}
