% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate2.R
\name{sexp2}
\alias{sexp2}
\title{Function that calls to estimate2 for a set of parameters}
\usage{
sexp2(x, param, rp, iteration)
}
\arguments{
\item{x}{stream to analyze}

\item{param}{parameters for the method of estimation}

\item{rp}{real value of p parameter}

\item{iteration}{iteration related to this experiment}
}
\description{
Function that calls to estimate2 for a set of parameters
}
