% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate3-FF.R
\name{sexp3}
\alias{sexp3}
\title{Function that calls to estimate3 for a set of parameters}
\usage{
sexp3(x, param, rp, iteration)
}
\arguments{
\item{x}{stream to analyze}

\item{param}{parameters for the method of estimation}

\item{rp}{real value of p parameter}

\item{iteration}{iteration related to this experiment}
}
\description{
Function that calls to estimate3 for a set of parameters
}
