% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate4-FW.R
\name{sexp4}
\alias{sexp4}
\title{Function that calls to estimate4 for a set of parameters}
\usage{
sexp4(x, param, rp, iteration)
}
\arguments{
\item{x}{stream to analyze}

\item{param}{parameters for the method of estimation}

\item{rp}{real value of p parameter}

\item{iteration}{iteration related to this experiment}
}
\description{
Function that calls to estimate4 for a set of parameters
}
