% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate5.R
\name{sexp5}
\alias{sexp5}
\title{Function that calls to estimate5 for a set of parameters}
\usage{
sexp5(x, param, rp, iteration)
}
\arguments{
\item{x}{stream to analyze}

\item{param}{parameters for the method of estimation}

\item{rp}{real value of p parameter}

\item{iteration}{iteration related to this experiment}
}
\description{
Function that calls to estimate5 for a set of parameters
}
