% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate8-SWF.R
\name{sexp8}
\alias{sexp8}
\title{Function that calls to estimate8 for a set of parameters}
\usage{
sexp8(x, param, rp, iteration)
}
\arguments{
\item{x}{stream to analyze}

\item{param}{parameters for the method of estimation}

\item{rp}{real value of p parameter}

\item{iteration}{iteration related to this experiment}
}
\description{
Function that calls to estimate8 for a set of parameters
}
