% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{simulate}
\alias{simulate}
\title{Simulate a series of binomial values}
\usage{
simulate(n, t)
}
\arguments{
\item{n}{the size of the sample for each probability value (the same for all)}

\item{t}{a vector of the different probability values}
}
\description{
Simulate a series of binomial values
}
