% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{test}
\alias{test}
\title{Function test with initial simulated data and probability estimations
It computes the average log likelihood of the observations with the
estimations of the previous step}
\usage{
test(x, y)
}
\arguments{
\item{x}{stream data to analyze}

\item{y}{estimation of the data}
}
\value{
average loglikelihood
}
\description{
Function test with initial simulated data and probability estimations
It computes the average log likelihood of the observations with the
estimations of the previous step
}
