% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{testel}
\alias{testel}
\title{Function that estimates probabilities according the
Bifet, Gavalda, 2007 procedure}
\usage{
testel(n1, n2, l1, l2, delta)
}
\arguments{
\item{n1}{number of 1s in the first sequence}

\item{n2}{number of 1s in the second sequence}

\item{l1}{length of the first sequence}

\item{l2}{length of the second sequence}

\item{delta}{parameter for test computation}
}
\value{
boolean flag
}
\description{
Function that estimates probabilities according the
Bifet, Gavalda, 2007 procedure
}
